setting up basic cache
