--- TroveQL/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "troveql",
-  "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "build": "tsc --watch"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@types/node": "^18.15.5",
-    "typescript": "^5.0.2"
-  },
-  "devDependencies": {
-    "ts-node": "^10.9.1"
-  }
-}
